[Test] isolate unit tests from `configurationHub.test.ts`
